#6 refac: refactoring directory
- 패키지를 notice -> subscribe로 변경했습니다.
