+++ ppap/src/main/java/com/ppap/ppap/domain/subscribe/dto/SubscribeCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.ppap.ppap.domain.notice.dto;
+package com.ppap.ppap.domain.subscribe.dto;
 
 import jakarta.validation.constraints.NotBlank;
