+++ ppap/src/main/java/com/ppap/ppap/domain/subscribe/repository/SubscribeRepository.java
@@ -1,7 +1,7 @@
-package com.ppap.ppap.domain.notice.repository;
+package com.ppap.ppap.domain.subscribe.repository;
 
-import com.ppap.ppap.domain.notice.entity.Notice;
-import com.ppap.ppap.domain.notice.entity.Subscribe;
+import com.ppap.ppap.domain.subscribe.entity.Notice;
+import com.ppap.ppap.domain.subscribe.entity.Subscribe;
 import com.ppap.ppap.domain.user.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
